import scenarios from excel
Instead of using expand.scenarios import scenarios from a csv file.
csv originates from xlsx that has formulas, thus if scenarios change
the xlsx must be saved and then also the first worksheet must be saved
as a csv. i might add a make file to the folder which would create the
csv any time the xlsx is updated.