Auto update scenarios.csv
scenarios.csv is no longer tracked within git, instead now we are tracking
just the xlsx file and a makefile will automatically convert the xlsx
file into the csv when runEstGrowth is called from R using a system call
this will help keep the scenario list up to date. Please let me know if
the makefile is not working on your computer as it is only tested on
windows machines.